refactor: code-review 指摘のうち simplify 対象 2 件を反映
- helper/rdb: RdbClient を D1 から導出する単一の真実に簡素化。冗長な
  setTestRdb と二重状態を撤去（setup は setTestD1 のみ呼ぶ）
- server/cron の vitest config: 重複していた pool-workers 設定
  (APP_ROOT/readD1Migrations/poolOptions/plugins/pool)を
  workers-pool.ts の createWorkersPool へ集約

correctness 指摘(setupFiles順序/isolatedStorage migrations/global-setup競合等)は
cron 9/9・server DBパス・CI server green で実証的に否定済みのため変更なし。
