+++ application/src/test/vitest/server/config.ts
@@ -1,38 +1,21 @@
 /// <reference types="vitest" />
-import { dirname, resolve } from 'node:path'
-import { fileURLToPath } from 'node:url'
-import { cloudflarePool, cloudflareTest, readD1Migrations } from '@cloudflare/vitest-pool-workers'
 import { defineConfig } from 'vitest/config'
 import { coverageReporter, generateIncludes } from '../generate'
+import { createWorkersPool } from '../workers-pool'
 
 const { testInclude, coverageInclude } = generateIncludes('src/web/server')
-const APP_ROOT = resolve(dirname(fileURLToPath(import.meta.url)), '..', '..', '..', '..')
 
 export default defineConfig(async () => {
-  // migrations を読み込み、各テスト(workerd内)の D1 へ applyD1Migrations で適用する
-  const migrations = await readD1Migrations(resolve(APP_ROOT, 'migrations'))
-
-  const poolOptions = {
-    miniflare: {
-      compatibilityDate: '2025-04-01',
-      compatibilityFlags: ['nodejs_compat'],
-      d1Databases: { DB: 'test-db' },
-      bindings: {
-        TEST_MIGRATIONS: migrations,
-        NODE_ENV: 'test',
-        LOG_LEVEL: 'silent',
-      },
-    },
-  }
+  const { plugins, pool } = await createWorkersPool()
 
   return {
-    plugins: [cloudflareTest(poolOptions)],
+    plugins,
     resolve: {
       tsconfigPaths: true,
     },
     test: {
       globals: true,
-      pool: cloudflarePool(poolOptions),
+      pool,
       setupFiles: ['src/test/setup/workers-d1.ts'],
       include: testInclude,
       coverage: {
